src/requirements.txt
Pillow==10.4.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.4
